// packages/core/src/types/generated/in-progress-events.ts
/* eslint-disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * Events from IM Step Function
 */
export type IMProgressEvent =
  | IMStarted
  | IMTrainingUnitCompleted
  | IMTrainingCompleted
  | IMEvaluationStarted
  | IMEvaluationCompleted
  | IMCompleted;

export interface IMStarted {
  event_type: "IMStarted";
  training_id: string;
  [k: string]: unknown;
}
export interface IMTrainingUnitCompleted {
  event_type: "IMTrainingUnitCompleted";
  training_id: string;
  completed_training_units: number;
  is_completed: boolean;
  [k: string]: unknown;
}
export interface IMTrainingCompleted {
  event_type: "IMTrainingCompleted";
  training_id: string;
  training_output: TrainingOutput;
  [k: string]: unknown;
}
/**
 * TrainingOutput contains the information about the training job.
 */
export interface TrainingOutput {
  /**
   * The S3 path where the training job saves the model artifacts.
   */
  output_path: string;
  /**
   * The number of training units completed.
   */
  completed_training_units: number;
  /**
   * Indicates whether the training job has completed, when all training units are  carried out
   */
  is_completed: boolean;
  [k: string]: unknown;
}
export interface IMEvaluationStarted {
  event_type: "IMEvaluationStarted";
  training_id: string;
  evaluation_id: string;
  [k: string]: unknown;
}
export interface IMEvaluationCompleted {
  event_type: "IMEvaluationCompleted";
  training_id: string;
  evaluation_output: EvaluationOutput;
  [k: string]: unknown;
}
/**
 * The output of a single evaluation run
 */
export interface EvaluationOutput {
  evaluation_id: string;
  /**
   * The path to the replay file
   */
  replay_path: string;
  /**
   * The stats for the evaluation interpretations
   */
  stats: (HistogramStats | AggregationStats)[];
  [k: string]: unknown;
}
/**
 * Represent a table data-structure for histogram statistics
 */
export interface HistogramStats {
  type: "histogram";
  /**
   * The name of the histogram
   */
  name: string;
  /**
   * The name of the x-axis
   */
  x_axis: string;
  /**
   * The name of the y-axis
   */
  y_axis: string;
  categories: string[];
  data: number[];
  [k: string]: unknown;
}
/**
 * Represent a aggregation of statistics.
 */
export interface AggregationStats {
  /**
   * The type of the aggregation.
   */
  type: "aggregation";
  /**
   * The name of the aggregation.
   */
  name: string;
  /**
   * The data of the aggregation.
   */
  data: number;
  [k: string]: unknown;
}
export interface IMCompleted {
  event_type: "IMCompleted";
  training_id: string;
  training_output: TrainingOutput;
  evaluation_output: EvaluationOutput[];
  [k: string]: unknown;
}